add mcompile command to build targets in a backend agnostic way
